+++ apps/server/src/console/console.module.ts
@@ -2,13 +2,12 @@ import { Module } from '@nestjs/common';
 import { ConsoleModule } from 'nestjs-console';
 import { ConsoleWriterModule } from '@shared/infra/console/console-writer/console-writer.module';
 import { ManagementModule } from '../modules/management/management.module';
-import { ConsoleWriter } from './console-writer';
+import { FilesModule } from '../modules/files/files.module';
 import { ServerConsole } from './server.console';
 
 @Module({
-	imports: [ManagementModule, ConsoleModule, ConsoleWriterModule],
+	imports: [ManagementModule, ConsoleModule, ConsoleWriterModule, FilesModule],
 	providers: [
-		ConsoleWriter,
 		/** add console services as providers */
 		ServerConsole,
 	],