+++ apps/server/src/modules/task/controller/dto/task.response.ts
@@ -1,17 +1,27 @@
 import { ApiProperty, ApiPropertyOptional } from '@nestjs/swagger';
+import { PaginationResponse, DecodeHtmlEntities } from '@shared/controller';
 
 /**
  * DTO for returning a task document via api.
  */
 export class TaskResponse {
 	@ApiProperty()
+	@DecodeHtmlEntities()
 	name: string;
+
+	@ApiProperty()
+	availableDate?: Date;
 
 	@ApiProperty()
 	duedate?: Date;
 
 	@ApiPropertyOptional()
+	@DecodeHtmlEntities()
 	courseName?: string;
+
+	@ApiPropertyOptional()
+	@DecodeHtmlEntities()
+	description?: string;
 
 	@ApiPropertyOptional()
 	displayColor?: string;
@@ -30,5 +40,16 @@ export class TaskResponse {
 		submitted?: number;
 		maxSubmissions?: number;
 		graded?: number;
+		isDraft: boolean;
 	};
+}
+
+export class TaskListResponse extends PaginationResponse<TaskResponse[]> {
+	constructor(data: TaskResponse[], total: number, skip?: number, limit?: number) {
+		super(total, skip, limit);
+		this.data = data;
+	}
+
+	@ApiProperty({ type: [TaskResponse] })
+	data: TaskResponse[];
 }