+++ apps/server/src/modules/task/mapper/task.mapper.ts
@@ -8,6 +8,7 @@ export class TaskMapper {
 
 		dto.id = task.id;
 		dto.name = task.name;
+		dto.availableDate = task.availableDate;
 		dto.duedate = task.dueDate;
 		dto.createdAt = task.createdAt;
 		dto.updatedAt = task.updatedAt;
@@ -15,13 +16,13 @@ export class TaskMapper {
 			submitted: status.submitted,
 			maxSubmissions: status.maxSubmissions,
 			graded: status.graded,
+			isDraft: status.isDraft,
 		};
 
-		if (task.parent !== undefined) {
-			const parentData = task.parent.getDescriptions();
-			dto.courseName = parentData.name;
-			dto.displayColor = parentData.color;
-		}
+		const taskDesc = task.getDescriptions();
+		dto.courseName = taskDesc.name;
+		dto.displayColor = taskDesc.color;
+		dto.description = taskDesc.description;
 
 		return dto;
 	}