+++ apps/server/src/modules/task/uc/task.uc.spec.ts
@@ -1,28 +1,27 @@
 import { Test, TestingModule } from '@nestjs/testing';
-
-import { createCurrentTestUser } from '@src/modules/user/utils';
-import { PaginationQuery } from '@shared/controller';
-
-import { EntityId, ICurrentUser, Submission, Task } from '@shared/domain';
 import { ObjectId } from '@mikro-orm/mongodb';
 import { Collection } from '@mikro-orm/core';
 
-import { userFactory } from '@shared/domain/factory';
+import { createCurrentTestUser } from '@src/modules/user/utils';
 import { MongoMemoryDatabaseModule } from '@src/modules/database';
-import { courseFactory } from '@shared/domain/factory/course.factory';
+import { userFactory, courseFactory } from '@shared/domain/factory';
+import { PaginationQuery } from '@shared/controller';
+import { EntityId, ICurrentUser, Lesson, Submission, Task } from '@shared/domain';
+import { LessonRepo } from '@shared/repo';
+
 import { TaskRepo } from '../repo';
 
 import { TaskUC, TaskDashBoardPermission } from './task.uc';
 import { TaskAuthorizationService, TaskParentPermission } from './task.authorization.service';
 
-// TODO: courseGroups test completly missing
-// TODO: how work this stuff with lessons
-// TODO: how work this stuff with ignoredTask
+// TODO: add courseGroups tests
+// TODO: what about ignoredTask?
 
-describe('TaskService', () => {
+describe('TaskUC', () => {
 	let module: TestingModule;
 	let service: TaskUC;
 	let taskRepo: TaskRepo;
+	let lessonRepo: LessonRepo;
 	let authorizationService: TaskAuthorizationService;
 
 	beforeEach(async () => {
@@ -33,11 +32,16 @@ describe('TaskService', () => {
 				{
 					provide: TaskRepo,
 					useValue: {
-						findAll() {
-							throw new Error('Please write a mock for TaskRepo.findAll');
+						findAllByParentIds() {
+							throw new Error('Please write a mock for TaskRepo.findAllByParentIds');
 						},
-						findAllCurrent() {
-							throw new Error('Please write a mock for TaskRepo.findAllCurrent');
+					},
+				},
+				{
+					provide: LessonRepo,
+					useValue: {
+						findAllByCourseIds() {
+							throw new Error('Please write a mock for LessonRepo.findAllByCourseIds');
 						},
 					},
 				},
@@ -54,6 +58,7 @@ describe('TaskService', () => {
 
 		service = module.get(TaskUC);
 		taskRepo = module.get(TaskRepo);
+		lessonRepo = module.get(LessonRepo);
 		authorizationService = module.get(TaskAuthorizationService);
 	});
 
@@ -62,24 +67,28 @@ describe('TaskService', () => {
 	});
 
 	const setTaskRepoMock = {
-		findAllCurrent: (data: Task[] = []) => {
-			const spy = jest.spyOn(taskRepo, 'findAllCurrent').mockImplementation(() => {
-				return Promise.resolve([data, data.length]);
+		findAllByParentIds: (tasks: Task[] = []) => {
+			const spy = jest.spyOn(taskRepo, 'findAllByParentIds').mockImplementation(() => {
+				return Promise.resolve([tasks, tasks.length]);
 			});
 			return spy;
 		},
-		findAll: (data: Task[] = []) => {
-			const spy = jest.spyOn(taskRepo, 'findAll').mockImplementation(() => {
-				return Promise.resolve([data, data.length]);
+	};
+
+	const setLessonRepoMock = {
+		findAllByCourseIds: (lessons: Lesson[] = []) => {
+			const spy = jest.spyOn(lessonRepo, 'findAllByCourseIds').mockImplementation(() => {
+				return Promise.resolve(lessons);
 			});
 			return spy;
 		},
 	};
 
 	const setAuthorizationServiceMock = {
-		getPermittedCourses: (data: EntityId[] = []) => {
+		// TODO: course instant of courseIds
+		getPermittedCourses: (courseIds: EntityId[] = []) => {
 			const spy = jest.spyOn(authorizationService, 'getPermittedCourses').mockImplementation(() => {
-				return Promise.resolve(data);
+				return Promise.resolve(courseIds);
 			});
 			return spy;
 		},
@@ -91,8 +100,8 @@ describe('TaskService', () => {
 
 	describe('findAll', () => {
 		const findAllMock = () => {
-			const spy1 = setTaskRepoMock.findAll();
-			const spy2 = setTaskRepoMock.findAllCurrent();
+			const spy1 = setTaskRepoMock.findAllByParentIds();
+			const spy2 = setLessonRepoMock.findAllByCourseIds();
 			const spy3 = setAuthorizationServiceMock.getPermittedCourses();
 
 			const mockRestore = () => {
@@ -149,104 +158,120 @@ describe('TaskService', () => {
 	describe('as a student', () => {
 		let currentUser: ICurrentUser;
 
+		const mockAll = (tasks?, lessons?, courseIds?) => {
+			const spy1 = setTaskRepoMock.findAllByParentIds(tasks);
+			const spy2 = setLessonRepoMock.findAllByCourseIds(lessons);
+			const spy3 = setAuthorizationServiceMock.getPermittedCourses(courseIds);
+
+			const mockRestore = () => {
+				spy1.mockRestore();
+				spy2.mockRestore();
+				spy3.mockRestore();
+			};
+
+			return mockRestore;
+		};
+
 		beforeEach(() => {
 			const permissions = [TaskDashBoardPermission.studentDashboard];
 			({ currentUser } = createCurrentTestUser(permissions));
 		});
 
 		it('should get parent ids for student role', async () => {
-			const spyTaskRepoFindAll = setTaskRepoMock.findAll([]);
-			const spyTaskRepoFindAllCurrent = setTaskRepoMock.findAllCurrent([]);
-			const spyGetPermittedCourses = setAuthorizationServiceMock.getPermittedCourses([]);
+			const mockRestore = mockAll();
+			const spy = setAuthorizationServiceMock.getPermittedCourses();
 
 			const paginationQuery = new PaginationQuery();
 			await service.findAll(currentUser, paginationQuery);
 
 			const expectedParams = [currentUser.userId, TaskParentPermission.read];
-			expect(spyGetPermittedCourses).toHaveBeenCalledWith(...expectedParams);
+			expect(spy).toHaveBeenCalledWith(...expectedParams);
 
-			spyTaskRepoFindAll.mockRestore();
-			spyTaskRepoFindAllCurrent.mockRestore();
-			spyGetPermittedCourses.mockRestore();
+			mockRestore();
 		});
 
 		it('should return a counted result', async () => {
-			const spyTaskRepoFindAll = setTaskRepoMock.findAll([]);
-			const spyTaskRepoFindAllCurrent = setTaskRepoMock.findAllCurrent([]);
-			const spyGetPermittedCourses = setAuthorizationServiceMock.getPermittedCourses([]);
+			const mockRestore = mockAll();
 
 			const paginationQuery = new PaginationQuery();
 			const [result, count] = await service.findAll(currentUser, paginationQuery);
 			expect(Array.isArray(result)).toBeTruthy();
 			expect(count).toEqual(0);
 
-			spyTaskRepoFindAll.mockRestore();
-			spyTaskRepoFindAllCurrent.mockRestore();
-			spyGetPermittedCourses.mockRestore();
+			mockRestore();
 		});
 
 		it('should find current tasks by permitted parent ids ordered by dueDate', async () => {
-			const spyTaskRepoFindAllCurrent = setTaskRepoMock.findAllCurrent([]);
+			const spy = setTaskRepoMock.findAllByParentIds([]);
+			const course = courseFactory.build();
+			const lesson = new Lesson({ name: 'lesson #1', course, hidden: false });
+			Object.assign(lesson, { _id: new ObjectId() });
+			const spyLessonRepoFindAllByCourseIds = setLessonRepoMock.findAllByCourseIds([lesson]);
 			const parentIds = [new ObjectId().toHexString(), new ObjectId().toHexString(), new ObjectId().toHexString()];
 			const spyGetPermittedCourses = setAuthorizationServiceMock.getPermittedCourses(parentIds);
 
 			const paginationQuery = new PaginationQuery();
 			await service.findAll(currentUser, paginationQuery);
 
-			const expectedParams = [
-				parentIds,
-				{ order: { dueDate: 'asc' }, pagination: { skip: paginationQuery.skip, limit: paginationQuery.limit } },
-			];
-			expect(spyTaskRepoFindAllCurrent).toHaveBeenCalledWith(...expectedParams);
+			expect(spy).toHaveBeenCalledTimes(1);
+			expect(spy.mock.calls[0][0]).toEqual({
+				courseIds: parentIds,
+				lessonIds: [lesson.id],
+			});
+			expect(spy.mock.calls[0][1]?.draft).toEqual(false);
+			expect(spy.mock.calls[0][1]?.closed).toEqual(currentUser.userId);
+			expect(spy.mock.calls[0][1]?.afterDueDateOrNone).toBeDefined();
+			expect(spy.mock.calls[0][2]).toEqual({
+				order: { dueDate: 'asc' },
+				pagination: { skip: paginationQuery.skip, limit: paginationQuery.limit },
+			});
 
-			spyTaskRepoFindAllCurrent.mockRestore();
+			expect(spyLessonRepoFindAllByCourseIds).toHaveBeenCalledWith(parentIds, { hidden: false });
+
+			spy.mockRestore();
+			spyLessonRepoFindAllByCourseIds.mockRestore();
 			spyGetPermittedCourses.mockRestore();
 		});
 
-		it('should return well formed task with parent and status', async () => {
+		it('should return well formed task with course and status', async () => {
 			const course = courseFactory.build();
-			const task = new Task({ name: 'task #1', private: false, parent: course });
+			const task = new Task({ name: 'task #1', private: false, course });
 
-			const spyTaskRepoFindAllCurrent = setTaskRepoMock.findAllCurrent([task]);
-			const spyGetPermittedCourses = setAuthorizationServiceMock.getPermittedCourses();
+			const mockRestore = mockAll([task]);
 
 			const paginationQuery = new PaginationQuery();
 			const [result] = await service.findAll(currentUser, paginationQuery);
-			expect(result[0]).toEqual({ task, status: { submitted: 0, maxSubmissions: 1, graded: 0 } });
-			expect(result[0].task.parent).toBeDefined();
+			expect(result[0]).toEqual({ task, status: { submitted: 0, maxSubmissions: 1, graded: 0, isDraft: false } });
+			expect(result[0].task.course).toBeDefined();
 
-			spyTaskRepoFindAllCurrent.mockRestore();
-			spyGetPermittedCourses.mockRestore();
+			mockRestore();
 		});
 
 		it('should find a list of tasks', async () => {
 			const course = courseFactory.build();
-			const task1 = new Task({ name: 'task #1', private: false, parent: course });
-			const task2 = new Task({ name: 'task #2', private: false, parent: course });
-			const task3 = new Task({ name: 'task #2', private: false, parent: course });
+			const task1 = new Task({ name: 'task #1', private: false, course });
+			const task2 = new Task({ name: 'task #2', private: false, course });
+			const task3 = new Task({ name: 'task #2', private: false, course });
 
-			const spyTaskRepoFindAllCurrent = setTaskRepoMock.findAllCurrent([task1, task2, task3]);
-			const spyGetPermittedCourses = setAuthorizationServiceMock.getPermittedCourses();
+			const mockRestore = mockAll([task1, task2, task3]);
 
 			const paginationQuery = new PaginationQuery();
 			const [result, count] = await service.findAll(currentUser, paginationQuery);
 			expect(count).toEqual(3);
 			expect(result.length).toEqual(3);
 
-			spyTaskRepoFindAllCurrent.mockRestore();
-			spyGetPermittedCourses.mockRestore();
+			mockRestore();
 		});
 
 		it('should compute submitted status for task', async () => {
 			const student = userFactory.build({ firstName: 'John', lastName: 'Doe' });
 			student.id = currentUser.userId;
 			const course = courseFactory.build();
-			const task = new Task({ name: 'task #1', private: false, parent: course });
+			const task = new Task({ name: 'task #1', private: false, course });
 			const submission = new Submission({ task, student, comment: 'my solution to the task #1' });
 			task.submissions = new Collection<Submission>(task, [submission]);
 
-			const spyTaskRepoFindAllCurrent = setTaskRepoMock.findAllCurrent([task]);
-			const spyGetPermittedCourses = setAuthorizationServiceMock.getPermittedCourses();
+			const mockRestore = mockAll([task]);
 
 			const paginationQuery = new PaginationQuery();
 			const [result] = await service.findAll(currentUser, paginationQuery);
@@ -256,24 +281,24 @@ describe('TaskService', () => {
 				graded: 0,
 				submitted: 1,
 				maxSubmissions: 1,
+				isDraft: false,
 			});
 
-			spyTaskRepoFindAllCurrent.mockRestore();
-			spyGetPermittedCourses.mockRestore();
+			mockRestore();
 		});
+
 		it('should only count the submissions of the given user', async () => {
 			const student1 = userFactory.build({ firstName: 'John', lastName: 'Doe' });
 			student1.id = currentUser.userId;
 			const student2 = userFactory.build({ firstName: 'Marla', lastName: 'Mathe' });
 			student2.id = new ObjectId().toHexString();
 			const course = courseFactory.build();
-			const task = new Task({ name: 'task #1', private: false, parent: course });
+			const task = new Task({ name: 'task #1', private: false, course });
 			const submission1 = new Submission({ task, student: student1, comment: 'submission #1' });
 			const submission2 = new Submission({ task, student: student2, comment: 'submission #2' });
 			task.submissions = new Collection<Submission>(task, [submission1, submission2]);
 
-			const spyTaskRepoFindAllCurrent = setTaskRepoMock.findAllCurrent([task]);
-			const spyGetPermittedCourses = setAuthorizationServiceMock.getPermittedCourses();
+			const mockRestore = mockAll([task]);
 
 			const paginationQuery = new PaginationQuery();
 			const [result] = await service.findAll(currentUser, paginationQuery);
@@ -283,23 +308,22 @@ describe('TaskService', () => {
 				graded: 0,
 				submitted: 1,
 				maxSubmissions: 1,
+				isDraft: false,
 			});
 
-			spyTaskRepoFindAllCurrent.mockRestore();
-			spyGetPermittedCourses.mockRestore();
+			mockRestore();
 		});
 
 		it('should compute graded status for task', async () => {
 			const student = userFactory.build({ firstName: 'John', lastName: 'Doe' });
 			student.id = currentUser.userId;
 			const course = courseFactory.build();
-			const task = new Task({ name: 'task #1', private: false, parent: course });
+			const task = new Task({ name: 'task #1', private: false, course });
 			const submission = new Submission({ task, student, comment: 'my solution to the task #1' });
 			task.submissions = new Collection<Submission>(task, [submission]);
 
 			const spyGraded = jest.spyOn(submission, 'isGraded').mockImplementation(() => true);
-			const spyTaskRepoFindAllCurrent = setTaskRepoMock.findAllCurrent([task]);
-			const spyGetPermittedCourses = setAuthorizationServiceMock.getPermittedCourses();
+			const mockRestore = mockAll([task]);
 
 			const paginationQuery = new PaginationQuery();
 			const [result] = await service.findAll(currentUser, paginationQuery);
@@ -310,10 +334,10 @@ describe('TaskService', () => {
 				graded: 1,
 				submitted: 1,
 				maxSubmissions: 1,
+				isDraft: false,
 			});
 
-			spyTaskRepoFindAllCurrent.mockRestore();
-			spyGetPermittedCourses.mockRestore();
+			mockRestore();
 			spyGraded.mockRestore();
 		});
 
@@ -323,15 +347,14 @@ describe('TaskService', () => {
 			const student2 = userFactory.build({ firstName: 'Marla', lastName: 'Mathe' });
 			student2.id = new ObjectId().toHexString();
 			const course = courseFactory.build();
-			const task = new Task({ name: 'task #1', private: false, parent: course });
+			const task = new Task({ name: 'task #1', private: false, course });
 			const submission1 = new Submission({ task, student: student1, comment: 'submission #1' });
 			const submission2 = new Submission({ task, student: student2, comment: 'submission #2' });
 			task.submissions = new Collection<Submission>(task, [submission1, submission2]);
 
 			jest.spyOn(submission1, 'isGraded').mockImplementation(() => true);
 			jest.spyOn(submission2, 'isGraded').mockImplementation(() => true);
-			const spyTaskRepoFindAllCurrent = setTaskRepoMock.findAllCurrent([task]);
-			const spyGetPermittedCourses = setAuthorizationServiceMock.getPermittedCourses();
+			const mockRestore = mockAll([task]);
 
 			const paginationQuery = new PaginationQuery();
 			const [result] = await service.findAll(currentUser, paginationQuery);
@@ -341,15 +364,29 @@ describe('TaskService', () => {
 				graded: 1,
 				submitted: 1,
 				maxSubmissions: 1,
+				isDraft: false,
 			});
 
-			spyTaskRepoFindAllCurrent.mockRestore();
-			spyGetPermittedCourses.mockRestore();
+			mockRestore();
 		});
 	});
 
 	describe('as a teacher', () => {
 		let currentUser: ICurrentUser;
+
+		const mockAll = (tasks?, lessons?, courseIds?) => {
+			const spy1 = setTaskRepoMock.findAllByParentIds(tasks);
+			const spy2 = setLessonRepoMock.findAllByCourseIds(lessons);
+			const spy3 = setAuthorizationServiceMock.getPermittedCourses(courseIds);
+
+			const mockRestore = () => {
+				spy1.mockRestore();
+				spy2.mockRestore();
+				spy3.mockRestore();
+			};
+
+			return mockRestore;
+		};
 
 		beforeEach(() => {
 			const permissions = [TaskDashBoardPermission.teacherDashboard];
@@ -357,8 +394,7 @@ describe('TaskService', () => {
 		});
 
 		it('should get parent ids for teacher role', async () => {
-			const spyTaskRepoFindAll = setTaskRepoMock.findAll([]);
-			const spyTaskRepoFindAllCurrent = setTaskRepoMock.findAllCurrent([]);
+			const mockRestore = mockAll();
 			const spyGetPermittedCourses = setAuthorizationServiceMock.getPermittedCourses([]);
 
 			const paginationQuery = new PaginationQuery();
@@ -367,91 +403,86 @@ describe('TaskService', () => {
 			const expectedParams = [currentUser.userId, TaskParentPermission.write];
 			expect(spyGetPermittedCourses).toHaveBeenCalledWith(...expectedParams);
 
-			spyTaskRepoFindAll.mockRestore();
-			spyTaskRepoFindAllCurrent.mockRestore();
+			mockRestore();
 			spyGetPermittedCourses.mockRestore();
 		});
 
 		it('should return a counted result', async () => {
-			const spyTaskRepoFindAll = setTaskRepoMock.findAll([]);
-			const spyTaskRepoFindAllCurrent = setTaskRepoMock.findAllCurrent([]);
-			const spyGetPermittedCourses = setAuthorizationServiceMock.getPermittedCourses([]);
+			const mockRestore = mockAll();
 
 			const paginationQuery = new PaginationQuery();
 			const [result, count] = await service.findAll(currentUser, paginationQuery);
 			expect(Array.isArray(result)).toBeTruthy();
 			expect(count).toEqual(0);
 
-			spyTaskRepoFindAll.mockRestore();
-			spyTaskRepoFindAllCurrent.mockRestore();
-			spyGetPermittedCourses.mockRestore();
+			mockRestore();
 		});
 
 		it('should find all tasks by permitted parent ids ordered by newest first', async () => {
-			const spyTaskRepoFindAll = setTaskRepoMock.findAll([]);
 			const parentIds = [new ObjectId().toHexString(), new ObjectId().toHexString(), new ObjectId().toHexString()];
-			const spyGetPermittedCourses = setAuthorizationServiceMock.getPermittedCourses(parentIds);
+			const course = courseFactory.build();
+			const lesson = new Lesson({ name: 'lesson #1', course, hidden: false });
+			Object.assign(lesson, { _id: new ObjectId() });
+			const tasks = [];
+			const mockRestore = mockAll(tasks, [lesson], parentIds);
+			const spy = setTaskRepoMock.findAllByParentIds(tasks);
 
 			const paginationQuery = new PaginationQuery();
 			await service.findAll(currentUser, paginationQuery);
 
 			const expectedParams = [
-				parentIds,
-				{ order: { createdAt: 'desc' }, pagination: { skip: paginationQuery.skip, limit: paginationQuery.limit } },
+				{ teacherId: currentUser.userId, courseIds: parentIds, lessonIds: [lesson.id] },
+				{ closed: currentUser.userId },
+				{ order: { dueDate: 'desc' }, pagination: { skip: paginationQuery.skip, limit: paginationQuery.limit } },
 			];
-			expect(spyTaskRepoFindAll).toHaveBeenCalledWith(...expectedParams);
 
-			spyTaskRepoFindAll.mockRestore();
-			spyGetPermittedCourses.mockRestore();
+			expect(spy).toHaveBeenCalledWith(...expectedParams);
+
+			mockRestore();
 		});
 
-		it('should return well formed task with parent and status', async () => {
+		it('should return well formed task with course and status', async () => {
 			const course = courseFactory.build();
-			const task = new Task({ name: 'task #1', private: false, parent: course });
+			const task = new Task({ name: 'task #1', private: true, course });
 
-			const spyTaskRepoFindAll = setTaskRepoMock.findAll([task]);
-			const spyGetPermittedCourses = setAuthorizationServiceMock.getPermittedCourses();
+			const mockRestore = mockAll([task]);
 
 			const paginationQuery = new PaginationQuery();
 			const [result] = await service.findAll(currentUser, paginationQuery);
 			expect(result[0]).toEqual({
 				task,
-				status: { submitted: 0, maxSubmissions: course.getNumberOfStudents(), graded: 0 },
+				status: { submitted: 0, maxSubmissions: course.getNumberOfStudents(), graded: 0, isDraft: true },
 			});
-			expect(result[0].task.parent).toBeDefined();
+			expect(result[0].task.course).toBeDefined();
 
-			spyTaskRepoFindAll.mockRestore();
-			spyGetPermittedCourses.mockRestore();
+			mockRestore();
 		});
 
 		it('should find a list of tasks', async () => {
 			const course = courseFactory.build();
-			const task1 = new Task({ name: 'task #1', private: false, parent: course });
-			const task2 = new Task({ name: 'task #2', private: false, parent: course });
-			const task3 = new Task({ name: 'task #2', private: false, parent: course });
+			const task1 = new Task({ name: 'task #1', private: false, course });
+			const task2 = new Task({ name: 'task #2', private: false, course });
+			const task3 = new Task({ name: 'task #2', private: false, course });
 
-			const spyTaskRepoFindAll = setTaskRepoMock.findAll([task1, task2, task3]);
-			const spyGetPermittedCourses = setAuthorizationServiceMock.getPermittedCourses();
+			const mockRestore = mockAll([task1, task2, task3]);
 
 			const paginationQuery = new PaginationQuery();
 			const [result, count] = await service.findAll(currentUser, paginationQuery);
 			expect(count).toEqual(3);
 			expect(result.length).toEqual(3);
 
-			spyTaskRepoFindAll.mockRestore();
-			spyGetPermittedCourses.mockRestore();
+			mockRestore();
 		});
 
 		it('should compute submitted status for task', async () => {
 			const student = userFactory.build({ firstName: 'John', lastName: 'Doe' });
 			student.id = new ObjectId().toHexString();
 			const course = courseFactory.build();
-			const task = new Task({ name: 'task #1', private: false, parent: course });
+			const task = new Task({ name: 'task #1', private: false, course });
 			const submission = new Submission({ task, student, comment: 'my solution to the task #1' });
 			task.submissions = new Collection<Submission>(task, [submission]);
 
-			const spyTaskRepoFindAll = setTaskRepoMock.findAll([task]);
-			const spyGetPermittedCourses = setAuthorizationServiceMock.getPermittedCourses();
+			const mockRestore = mockAll([task]);
 
 			const paginationQuery = new PaginationQuery();
 			const [result] = await service.findAll(currentUser, paginationQuery);
@@ -461,10 +492,10 @@ describe('TaskService', () => {
 				graded: 0,
 				submitted: 1,
 				maxSubmissions: course.getNumberOfStudents(),
+				isDraft: false,
 			});
 
-			spyTaskRepoFindAll.mockRestore();
-			spyGetPermittedCourses.mockRestore();
+			mockRestore();
 		});
 
 		it('should count all student ids of submissions', async () => {
@@ -473,13 +504,12 @@ describe('TaskService', () => {
 			const student2 = userFactory.build({ firstName: 'Marla', lastName: 'Mathe' });
 			student2.id = new ObjectId().toHexString();
 			const course = courseFactory.build();
-			const task = new Task({ name: 'task #1', private: false, parent: course });
+			const task = new Task({ name: 'task #1', private: false, course });
 			const submission1 = new Submission({ task, student: student1, comment: 'submission #1' });
 			const submission2 = new Submission({ task, student: student2, comment: 'submission #2' });
 			task.submissions = new Collection<Submission>(task, [submission1, submission2]);
 
-			const spyTaskRepoFindAll = setTaskRepoMock.findAll([task]);
-			const spyGetPermittedCourses = setAuthorizationServiceMock.getPermittedCourses();
+			const mockRestore = mockAll([task]);
 
 			const paginationQuery = new PaginationQuery();
 			const [result] = await service.findAll(currentUser, paginationQuery);
@@ -489,23 +519,22 @@ describe('TaskService', () => {
 				graded: 0,
 				submitted: 2,
 				maxSubmissions: course.getNumberOfStudents(),
+				isDraft: false,
 			});
 
-			spyTaskRepoFindAll.mockRestore();
-			spyGetPermittedCourses.mockRestore();
+			mockRestore();
 		});
 
 		it('should compute graded status for task', async () => {
 			const student = userFactory.build({ firstName: 'John', lastName: 'Doe' });
 			student.id = new ObjectId().toHexString();
 			const course = courseFactory.build();
-			const task = new Task({ name: 'task #1', private: false, parent: course });
+			const task = new Task({ name: 'task #1', private: false, course });
 			const submission = new Submission({ task, student, comment: 'my solution to the task #1' });
 			task.submissions = new Collection<Submission>(task, [submission]);
 
 			const spyGraded = jest.spyOn(submission, 'isGraded').mockImplementation(() => true);
-			const spyTaskRepoFindAll = setTaskRepoMock.findAll([task]);
-			const spyGetPermittedCourses = setAuthorizationServiceMock.getPermittedCourses();
+			const mockRestore = mockAll([task]);
 
 			const paginationQuery = new PaginationQuery();
 			const [result] = await service.findAll(currentUser, paginationQuery);
@@ -516,10 +545,10 @@ describe('TaskService', () => {
 				graded: 1,
 				submitted: 1,
 				maxSubmissions: course.getNumberOfStudents(),
+				isDraft: false,
 			});
 
-			spyTaskRepoFindAll.mockRestore();
-			spyGetPermittedCourses.mockRestore();
+			mockRestore();
 			spyGraded.mockRestore();
 		});
 
@@ -529,15 +558,14 @@ describe('TaskService', () => {
 			const student2 = userFactory.build({ firstName: 'Marla', lastName: 'Mathe' });
 			student2.id = new ObjectId().toHexString();
 			const course = courseFactory.build();
-			const task = new Task({ name: 'task #1', private: false, parent: course });
+			const task = new Task({ name: 'task #1', private: false, course });
 			const submission1 = new Submission({ task, student: student1, comment: 'submission #1' });
 			const submission2 = new Submission({ task, student: student2, comment: 'submission #2' });
 			task.submissions = new Collection<Submission>(task, [submission1, submission2]);
 
 			jest.spyOn(submission1, 'isGraded').mockImplementation(() => true);
 			jest.spyOn(submission2, 'isGraded').mockImplementation(() => true);
-			const spyTaskRepoFindAll = setTaskRepoMock.findAll([task]);
-			const spyGetPermittedCourses = setAuthorizationServiceMock.getPermittedCourses();
+			const mockRestore = mockAll([task]);
 
 			const paginationQuery = new PaginationQuery();
 			const [result] = await service.findAll(currentUser, paginationQuery);
@@ -547,10 +575,10 @@ describe('TaskService', () => {
 				graded: 2,
 				submitted: 2,
 				maxSubmissions: course.getNumberOfStudents(),
+				isDraft: false,
 			});
 
-			spyTaskRepoFindAll.mockRestore();
-			spyGetPermittedCourses.mockRestore();
+			mockRestore();
 		});
 
 		it('should count only unique student ids of graded submissions', async () => {
@@ -559,7 +587,7 @@ describe('TaskService', () => {
 			const student2 = userFactory.build({ firstName: 'Marla', lastName: 'Mathe' });
 			student2.id = new ObjectId().toHexString();
 			const course = courseFactory.build();
-			const task = new Task({ name: 'task #1', private: false, parent: course });
+			const task = new Task({ name: 'task #1', private: false, course });
 			const submission1 = new Submission({ task, student: student1, comment: 'submission #1' });
 			const submission2 = new Submission({ task, student: student2, comment: 'submission #2' });
 			const submission3 = new Submission({ task, student: student2, comment: 'submission #3' });
@@ -568,8 +596,7 @@ describe('TaskService', () => {
 			jest.spyOn(submission1, 'isGraded').mockImplementation(() => true);
 			jest.spyOn(submission2, 'isGraded').mockImplementation(() => true);
 			jest.spyOn(submission3, 'isGraded').mockImplementation(() => true);
-			const spyTaskRepoFindAll = setTaskRepoMock.findAll([task]);
-			const spyGetPermittedCourses = setAuthorizationServiceMock.getPermittedCourses();
+			const mockRestore = mockAll([task]);
 
 			const paginationQuery = new PaginationQuery();
 			const [result] = await service.findAll(currentUser, paginationQuery);
@@ -579,10 +606,10 @@ describe('TaskService', () => {
 				graded: 2,
 				submitted: 2,
 				maxSubmissions: course.getNumberOfStudents(),
+				isDraft: false,
 			});
 
-			spyTaskRepoFindAll.mockRestore();
-			spyGetPermittedCourses.mockRestore();
+			mockRestore();
 		});
 
 		it('should count only unique student ids of submissions', async () => {
@@ -591,14 +618,13 @@ describe('TaskService', () => {
 			const student2 = userFactory.build({ firstName: 'Marla', lastName: 'Mathe' });
 			student2.id = new ObjectId().toHexString();
 			const course = courseFactory.build();
-			const task = new Task({ name: 'task #1', private: false, parent: course });
+			const task = new Task({ name: 'task #1', private: false, course });
 			const submission1 = new Submission({ task, student: student1, comment: 'submission #1' });
 			const submission2 = new Submission({ task, student: student1, comment: 'submission #2' });
 			const submission3 = new Submission({ task, student: student2, comment: 'submission #2' });
 			task.submissions = new Collection<Submission>(task, [submission1, submission2, submission3]);
 
-			const spyTaskRepoFindAll = setTaskRepoMock.findAll([task]);
-			const spyGetPermittedCourses = setAuthorizationServiceMock.getPermittedCourses();
+			const mockRestore = mockAll([task]);
 
 			const paginationQuery = new PaginationQuery();
 			const [result] = await service.findAll(currentUser, paginationQuery);
@@ -608,10 +634,10 @@ describe('TaskService', () => {
 				graded: 0,
 				submitted: 2,
 				maxSubmissions: course.getNumberOfStudents(),
+				isDraft: false,
 			});
 
-			spyTaskRepoFindAll.mockRestore();
-			spyGetPermittedCourses.mockRestore();
+			mockRestore();
 		});
 	});
 });