+++ apps/server/src/shared/domain/entity/submission.entity.spec.ts
@@ -1,7 +1,6 @@
 import { Collection } from '@mikro-orm/core';
-import { Test, TestingModule } from '@nestjs/testing';
-import { MongoMemoryDatabaseModule } from '@src/modules/database';
-import { userFactory } from '../factory';
+import { setupEntities } from '@src/modules/database';
+import { userFactory, fileFactory } from '../factory';
 import { File } from './file.entity';
 import { Submission } from './submission.entity';
 import { Task } from './task.entity';
@@ -14,14 +13,8 @@ const buildSubmission = () => {
 };
 
 describe('Submission entity', () => {
-	let module: TestingModule;
-
 	beforeAll(async () => {
-		module = await Test.createTestingModule({ imports: [MongoMemoryDatabaseModule.forRoot()] }).compile();
-	});
-
-	afterAll(async () => {
-		await module.close();
+		await setupEntities();
 	});
 
 	it('should be graded if grade percentage is set', () => {
@@ -39,7 +32,7 @@ describe('Submission entity', () => {
 	it('should be graded if grade grade files have been associated', () => {
 		const submission = buildSubmission();
 		const teacher = userFactory.build({ firstName: 'Carl', lastName: 'Cord' });
-		const file = new File({ name: 'grade file', creator: teacher });
+		const file = fileFactory.build({ creator: teacher });
 		submission.gradeFiles = new Collection<File>(submission, [file]);
 		expect(submission.isGraded()).toEqual(true);
 	});