+++ apps/server/src/shared/domain/entity/task.entity.ts
@@ -1,33 +1,23 @@
-import { Collection, Entity, ManyToOne, OneToMany, Property } from '@mikro-orm/core';
+import { Collection, Entity, ManyToOne, OneToMany, ManyToMany, Property } from '@mikro-orm/core';
 import { BaseEntityWithTimestamps } from './base.entity';
-import { EntityId } from '../types';
 import type { Course } from './course.entity';
 import type { Lesson } from './lesson.entity';
 import type { Submission } from './submission.entity';
+import { User } from './user.entity';
 
 interface ITaskProperties {
 	name: string;
+	availableDate?: Date;
 	dueDate?: Date;
 	private?: boolean;
-	parent?: Course;
+	teacher?: User;
+	course?: Course;
 	lesson?: Lesson;
 	submissions?: Submission[];
+	closed?: User[];
 }
 
-export interface IParentDescriptionsProperties {
-	id: EntityId;
-	name: string;
-	color: string;
-	description?: string;
-}
-
-export interface ITaskParent {
-	id: EntityId;
-
-	hasWritePermission(userId: EntityId): boolean;
-	getDescriptions(): IParentDescriptionsProperties;
-	getNumberOfStudents(): number;
-}
+export type TaskParentDescriptions = { name: string; description: string; color: string };
 
 @Entity({ tableName: 'homeworks' })
 export class Task extends BaseEntityWithTimestamps {
@@ -35,13 +25,19 @@ export class Task extends BaseEntityWithTimestamps {
 	name: string;
 
 	@Property()
+	availableDate?: Date;
+
+	@Property()
 	dueDate?: Date;
 
 	@Property()
 	private = true;
 
+	@ManyToOne('User', { fieldName: 'teacherId' })
+	teacher?: User;
+
 	@ManyToOne('Course', { fieldName: 'courseId' })
-	parent?: Course;
+	course?: Course;
 
 	@ManyToOne('Lesson', { fieldName: 'lessonId' })
 	lesson?: Lesson; // In database exist also null, but it can not set.
@@ -49,13 +45,44 @@ export class Task extends BaseEntityWithTimestamps {
 	@OneToMany('Submission', 'task')
 	submissions = new Collection<Submission>(this);
 
+	// TODO: is mapped to boolean in future
+	@ManyToMany('User', undefined, { fieldName: 'archived' })
+	closed = new Collection<User>(this);
+
+	isDraft(): boolean {
+		// private can be undefined in the database
+		return !!this.private;
+	}
+
+	getDescriptions(): TaskParentDescriptions {
+		let descriptions: TaskParentDescriptions;
+		if (this.course) {
+			descriptions = {
+				name: this.course.name,
+				description: this.lesson ? this.lesson.name : '',
+				color: this.course.color,
+			};
+		} else {
+			descriptions = {
+				name: '',
+				description: '',
+				color: '#ACACAC',
+			};
+		}
+		return descriptions;
+	}
+
 	constructor(props: ITaskProperties) {
 		super();
 		this.name = props.name;
+		this.availableDate = props.availableDate;
 		this.dueDate = props.dueDate;
 		if (props.private !== undefined) this.private = props.private;
-		this.parent = props.parent;
+		this.teacher = props.teacher;
+		this.course = props.course;
 		this.lesson = props.lesson;
 		this.submissions.set(props.submissions || []);
+		// TODO: is replaced with boolean in future
+		this.closed.set(props.closed || []);
 	}
 }