+++ src/services/ldap/strategies/univention.js
@@ -13,11 +13,13 @@ class UniventionLDAPStrategy extends AbstractLDAPStrategy {
 	 */
 	getSchools() {
 		let ignoredSchools = '';
-		this.config.providerOptions.ignoreSchools.forEach((schoolOu) => {
-			ignoredSchools += `(!(ou=${schoolOu}))`;
-		});
+		if (Array.isArray(this.config.providerOptions.ignoreSchools)) {
+			this.config.providerOptions.ignoreSchools.forEach((schoolOu) => {
+				ignoredSchools += `(!(ou=${schoolOu}))`;
+			});
+		}
 		const options = {
-			filter: `(&(univentionObjectType=container/ou)(!(ucsschoolRole=school:ou:no_school))${ignoredSchools})`,
+			filter: `(&(univentionObjectType=container/ou)(!(ucsschoolRole=school:ou:no_school))(objectClass=ucsschoolOrganizationalUnit)${ignoredSchools})`,
 			scope: 'sub',
 			attributes: [],
 		};